added current track to q
