+++ components/Spaces/CallTabs/Music/Player.js
@@ -4,7 +4,7 @@ import SpotifyPlayer from 'react-spotify-web-playback';
 import { useSpotify } from './SpotifyProvider';
 
 const Player = () => {
-  const { accessToken, trackUri } = useSpotify();
+  const { accessToken, trackUri, setCurrentTrack, setNextTracks } = useSpotify();
   const [play, setPlay] = useState(false);
 
   useEffect(() => setPlay(true), [trackUri]);
@@ -16,6 +16,13 @@ const Player = () => {
       showSaveIcon
       callback={(state) => {
         if (!state.isPlaying) setPlay(false);
+        setCurrentTrack({
+          artist: state?.track?.artists,
+          title: state?.track?.name,
+          uri: state?.track?.uri,
+          albumUrl: state?.track?.image,
+        });
+        setNextTracks(state.nextTracks ?? []);
       }}
       play={play}
       uris={trackUri ? [trackUri] : []}
